Split testcases for different interface impl
